+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Karpinets Alexander on 29.01.2023.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let question: String
+    let correctAnswer: Bool
+}
