+++ MovieQuiz/Models/QuizResultViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultViewModel.swift
+//  MovieQuiz
+//
+//  Created by Karpinets Alexander on 29.01.2023.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
